requirements.txt

# Machine Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics==8.2.10
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
filterpy>=1.4.5  # For Kalman Filter tracking (ByteTrack / DeepSORT dependency)

# Graph Neural Networks
torch-geometric>=2.3.0

# API & Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.3
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0.1
pydantic>=2.0.0
